Do wiadomosci commita:
It was decided that we are not going to update encryption flag of tables
inside a tablespace (after it was encrypted) - as we did for keyring
encryption in 5.7. Instead we go with 8.0 upstream general tablespace
encryption design and update only encryption flag of
tablespace. Because of that the part of innodb_aborted_flushed which was
testing failure scenario when update of encryption flag of a table
inside a tablespace fails was removed.

do wiadomosci wlasnej:
crypt_data jest sprawdzane dla undo tablespace wczesniej niz czytane
sa dane z headera. Dzieki temu nie mam komunikatu o niewlasiwej wesji
musze to dodac do brancha z dd update.