PS-5547: redo log encryption is not working.
The encryption was disabled for all logs. This problem was overlooked as
upstream MTR tests does not GREP the redo log for plaintext, only check
for encryption markers. Thus although the upstream tests passed, the
redo itself was not encrypted.

The tests for GREPing the redo log will be commited as part of PS-5476.