PS-6850: upgrade_crypt_data_v1 is failing
upgrade_crypt_data_v1 was failing due to warning:

[InnoDB] Table encryption flag is OFF in the data dictionary
but the encryption flag in file ./test/t3.ibd is ON. This indicates that
the rotation of the table was interrupte
d before space's flags were updated. Please have encryption_thread
variable (innodb-encryption-threads) set to va
lue > 0. So the encryption could finish up the rotation.

This is bacause 8.0.16 data directory was created without full flush of
changes. This is why we need to ignore this warning.