PS-7140: crypt redo logs are not applied correctly
The redo log recovery works in two stages:
 - parsing stage
 - apply stage

During parsing stage we parse the log and add them into hash table. The
stact trace for parsing is:
-> recv_multi_rec
    -> recv_parse_log_rec();
       -> recv_parse_or_apply_log_rec_body(block == nullptr) -> thus no apply
    -> recv_parse_log_rec();
       -> recv_parse_or_apply_log_rec_body(block==nullptr) -> thus no apply
    -> recv_add_to_hash_table

The recv_parse_or_apply_log_rec_body is called twice. Its task here is
to parse the logs, without applying them. This is because block passed
to this function is equal to nullptr. Each log record is added to
hash_table.

After all logs have been parsed we switch to apply step. The stack trace
for it is:

recv_apply_hashed_log_recs
  -> fil_tablespace_open_for_recovery
  -> recv_apply_log_rec
     -> block = buf_page_get
     -> recv_recover_page(false, block)
       -> recv_parse_or_apply_log_rec_body (block != nullptr)

Each tablespace is opened for recovery. Parsed logs are this time
applied to pages (recv_parse_or_apply_log_rec_body is called with block
!= nullptr).

In our case, when we have create table statement only in redo and thus
write to page with crypt_data is also only in redo - during parsing
stage we cannot assign crypt_data to space, because it has not yet been
created (redo of create table has not yet been applied). Because of that
during parsing stage we harvest all the crypt_data and add them into
recv_sys->crypt_datas map. Then after tablespace gets created and we
open it for recovery we assign crypt_data to it.